Add files via upload
Here we present the collection of epithelial transport models from Weinstein et
al., from which we derive a consistent, reproducible, and reusable mathematical
model of proximal tubule transport.
The mathematical model of epithelial proximal convoluted tubule encoded
in Python describes the cellular and subcellular mechanism of the transporters, buffer
pairs, solute fluxes. In the current epithelial model, 35 nonlinear equations are solved using a finite-difference numerical method along the Python solver ’scipy.optimize.root’. The evaluation of the model involves integrating the mass conservation equations from an initial time to a final time using small time increments. We need to be sure that we run the epithelial model long enough to reach the steady-state and stay there for a sufficient amount of time. 
To reproduce the figures from the document, we run all the simulations for 2000 S with the time step of 0.1 S.